examples/hello_pyramid/requirements.txt
pyramid==1.10.4
pyramid-jinja2==2.8
pyramid-htmldoom==0.1
